extract spec_field_parse
